+++ src/lightly_train/_visualize/object_detection.py
@@ -51,7 +51,7 @@ def plot_object_detection_labels(
 
     pil_images: list[PILImage] = []
     for i in range(n):
-        image_tensor = gt_images[i].clone()
+        image_tensor = gt_images[i].clone().to(dtype=torch.float32)
         if mean is not None and std is not None:
             image_tensor = _denormalize_image(image=image_tensor, mean=mean, std=std)
 
@@ -123,7 +123,7 @@ def plot_object_detection_predictions(
 
     pil_images: list[PILImage] = []
     for i in range(n):
-        image_tensor = gt_images[i].clone()
+        image_tensor = gt_images[i].clone().to(dtype=torch.float32)
         if mean is not None and std is not None:
             image_tensor = _denormalize_image(
                 image=image_tensor,
